Update ridenbaughhall.md
This is sort of a test run-editing/commiting/etc. I am still working on what sort of final length/layout options might be good, and would love feedback.  Additionally- I would like to cite the source for this text, as it was edited from the nomination document for NRHP, and written by Nathan J. Moody. (I am not sure how to do citations/sources in this system)